pull_request.rs: Parse JSON payload to create `CommitRef` struct
The test now passes, and we create our `CommitRef` struct using actual
values from the parsed JSON.

In order to get the branch name, we need the part after the last "/" in
the "ref" key.

There are a ton of `unwrap`s here, but at least it's a start and it's
working now.