jenkins.rs: Make `get_jobs` real
Instead of just making a request to test out 'reqwest' and how to use
it, make this function actually do something useful.

It now requests a URL dynamically based on the `repo_name` passed in,
and returns a list of build job URLs from the JSON resulting from the
API response.