github.rs: `update_commit_status` should take a repo name
Pass a repo name explicitly in an argument to the function. We now have
the values we need to be able to properly construct a URL to POST to.

Update our HTTP request to send the params as JSON and include the
"Accept" header that GitHub recommends adding
(https://developer.github.com/v3/#current-version) based on the code
examples in
https://docs.rs/reqwest/0.8.1/reqwest/header/struct.Accept.html#examples.