main(): Add command line options
Add options for:

* Jenkins URL
* Jenkins username
* Jenkins token
* GitHub token
* Port

Add a help option for usage information. All options except "port" are
required. Didn't use `reqopt` because that panics when you try to ask
for help without the required options. Now, when trying to run without a
required option, the usage help is printed. Maybe we should have an
error message instead.