Take Jenkins URL from command line options
Get rid of the static, global `API_URL`, and replace it with the Jenkins
URL that gets passed into `find_and_track_build_and_update_status()`.

This necessitated adding a new URL argument to `get_jobs()` and
`request_job()`. Starting to wonder if I should be creating a struct to
hold the URL and client that I can attach these functions to as methods,
but for now I'll leave them alone.