request_job(): Use `Job::new()`
Realised that the reason why I was getting a dead code warning from
`Job::new` was not just because I wasn't using it, but because I was
meant to use it somewhere. Remove the duplicate work going on in
`request_job()` and replace it with a call to `Job::new()` instead.