find_and_track_build_and_update_status(): Try to wrap the fn in a thread
Try to put the whole function inside a thread. This doesn't work because
of move problems. The reason for wrapping things in a thread is so I can
add a `sleep` at the top of the function.

We want to sleep before even making our first request to Jenkins because
it seems like we're making the request too early, before the job has
been created. This prevents the GitHub status from being set.