main(): Add comment above cloned command line variables
Turns out the easiest way to share these variables between both closures
really is to clone them.

Thanks to 'durka42' on Mozilla#rust-beginners for explaining the problem
to me. In durka's words:

> both closures are required to be 'static
> so you can't send references across
> and the router handler has to be callable multiple times so you can't consume anything (that's the error you got)

It's also possible to share references using Arc, but that seems more
complicated than it's worth here. Just do the copy and be done with it.