pull_request.rs: Add `pull_request_opened_or_synchronized()`
A new function that, given the parsed JSON from a pull request webhook,
tells us whether the pull request was "opened" or "synchronize"d with a
boolean.

This will enable us to only update commit statuses when actions on the
pull request result in new commits being available. Otherwise, we don't
want to set a commit status that we've already set before.