main(): Don't update commit status unless commits were pushed
There are a bunch of events that can trigger the `pull_request` webhook.
These are:

    "assigned", "unassigned", "review_requested",
    "review_request_removed", "labeled", "unlabeled", "opened",
    "edited", "closed", or "reopened". And "synchronize".

We only care about "opened" and "synchronize" because those are the only
two where new commits can come through. Since we don't set commit
statuses unless that commit is part of a pull request, we need a way to
update the status the first time a pull request is opened, thus the
"opened" action. When new commits are added we also want to update
statuses, so "synchronize". But for the others, we don't want to be
duplicating work and adding duplicate unnecessary statuses to PR
commits.