find_and_track_build_and_update_status(): Send Jenkins console URL
When adding a GitHub commit status, instead of using the job "home" URL,
use the URL of its "Console" page. This page shows the build log and is
much more useful than the other one. If there's a build failure in
particular, you'd be going to this page anyway to see what the problem
was, so using this `target_url` saves a click.